+++ cohere/responses/chat.py
@@ -1,5 +1,6 @@
 import json
-from typing import Any, Dict, Generator, List, Optional
+from enum import Enum
+from typing import Any, Dict, Generator, List, Optional, Union
 
 import requests
 
@@ -19,6 +20,11 @@ class Chat(CohereObject):
         chatlog: Optional[List[Dict[str, str]]] = None,
         preamble: Optional[str] = None,
         token_count: Optional[Dict[str, int]] = None,
+        is_search_required: Optional[bool] = None,
+        citations: Optional[List[Dict[str, Any]]] = None,
+        documents: Optional[List[Dict[str, Any]]] = None,
+        search_results: Optional[List[Dict[str, Any]]] = None,
+        search_queries: Optional[List[Dict[str, Any]]] = None,
         client=None,
         **kwargs,
     ) -> None:
@@ -34,13 +40,18 @@ class Chat(CohereObject):
         self.client = client
         self.token_count = token_count
         self.meta = meta
+        self.is_search_required = is_search_required  # optional
+        self.citations = citations  # optional
+        self.documents = documents  # optional
+        self.search_results = search_results  # optional
+        self.search_queries = search_queries  # optional
 
     @classmethod
     def from_dict(cls, response: Dict[str, Any], message: str, client) -> "Chat":
         return cls(
             id=response["response_id"],
             response_id=response["response_id"],
-            generation_id=response["generation_id"],
+            generation_id=response.get("generation_id"),  # optional
             message=message,
             conversation_id=response.get("conversation_id"),  # optional
             text=response.get("text"),
@@ -50,6 +61,11 @@ class Chat(CohereObject):
             client=client,
             token_count=response.get("token_count"),
             meta=response.get("meta"),
+            is_search_required=response.get("is_search_required"),  # optional
+            citations=response.get("citations"),  # optional
+            documents=response.get("documents"),  # optional
+            search_results=response.get("search_results"),  # optional
+            search_queries=response.get("search_queries"),  # optional
         )
 
     def respond(self, response: str, max_tokens: int = None) -> "Chat":
@@ -75,16 +91,27 @@ class AsyncChat(Chat):
         )
 
 
+class StreamEvent(str, Enum):
+    STREAM_START = "stream-start"
+    SEARCH_QUERIES_GENERATION = "search-queries-generation"
+    SEARCH_RESULTS = "search-results"
+    TEXT_GENERATION = "text-generation"
+    CITATION_GENERATION = "citation-generation"
+    STREAM_END = "stream-end"
+
+
 class StreamResponse(CohereObject):
     def __init__(
         self,
         is_finished: bool,
+        event_type: Union[StreamEvent, str],
         index: Optional[int],
         **kwargs,
     ) -> None:
         super().__init__(**kwargs)
         self.is_finished = is_finished
         self.index = index
+        self.event_type = event_type
 
 
 class StreamStart(StreamResponse):
@@ -109,6 +136,48 @@ class StreamTextGeneration(StreamResponse):
         self.text = text
 
 
+class StreamCitationGeneration(StreamResponse):
+    def __init__(
+        self,
+        citations: Optional[List[Dict[str, Any]]],
+        **kwargs,
+    ) -> None:
+        super().__init__(**kwargs)
+        self.citations = citations
+
+
+class StreamQueryGeneration(StreamResponse):
+    def __init__(
+        self,
+        search_queries: Optional[List[Dict[str, Any]]],
+        **kwargs,
+    ) -> None:
+        super().__init__(**kwargs)
+        self.search_queries = search_queries
+
+
+class StreamSearchResults(StreamResponse):
+    def __init__(
+        self,
+        search_results: Optional[List[Dict[str, Any]]],
+        documents: Optional[List[Dict[str, Any]]],
+        **kwargs,
+    ) -> None:
+        super().__init__(**kwargs)
+        self.search_results = search_results
+        self.documents = documents
+
+
+class StreamEnd(StreamResponse):
+    def __init__(
+        self,
+        finish_reason: str,
+        **kwargs,
+    ) -> None:
+        super().__init__(**kwargs)
+        self.finish_reason = finish_reason
+
+
 class StreamingChat(CohereObject):
     def __init__(self, response):
         self.response = response
@@ -122,23 +191,51 @@ class StreamingChat(CohereObject):
         self.finish_reason = None
         self.token_count = None
         self.meta = None
+        self.is_search_required = None
+        self.citations = None
+        self.documents = None
+        self.search_results = None
+        self.search_queries = None
 
     def _make_response_item(self, index, line) -> Any:
         streaming_item = json.loads(line)
         event_type = streaming_item.get("event_type")
 
-        if event_type == "stream-start":
+        if event_type == StreamEvent.STREAM_START:
             self.conversation_id = streaming_item.get("conversation_id")
             self.generation_id = streaming_item.get("generation_id")
             return StreamStart(
-                conversation_id=self.conversation_id, generation_id=self.generation_id, is_finished=False, index=index
+                conversation_id=self.conversation_id,
+                generation_id=self.generation_id,
+                is_finished=False,
+                event_type=event_type,
+                index=index,
             )
-        elif event_type == "text-generation":
+        elif event_type == StreamEvent.SEARCH_QUERIES_GENERATION:
+            search_queries = streaming_item.get("search_queries")
+            return StreamQueryGeneration(
+                search_queries=search_queries, is_finished=False, event_type=event_type, index=index
+            )
+        elif event_type == StreamEvent.SEARCH_RESULTS:
+            search_results = streaming_item.get("search_results")
+            documents = streaming_item.get("documents")
+            return StreamSearchResults(
+                search_results=search_results,
+                documents=documents,
+                is_finished=False,
+                event_type=event_type,
+                index=index,
+            )
+        elif event_type == StreamEvent.TEXT_GENERATION:
             text = streaming_item.get("text")
-            return StreamTextGeneration(text=text, is_finished=False, index=index)
-        elif event_type == "stream-end":
+            return StreamTextGeneration(text=text, is_finished=False, event_type=event_type, index=index)
+        elif event_type == StreamEvent.CITATION_GENERATION:
+            citations = streaming_item.get("citations")
+            return StreamCitationGeneration(citations=citations, is_finished=False, event_type=event_type, index=index)
+        elif event_type == StreamEvent.STREAM_END:
             response = streaming_item.get("response")
-            self.finish_reason = streaming_item.get("finish_reason")
+            finish_reason = streaming_item.get("finish_reason")
+            self.finish_reason = finish_reason
 
             if response is None:
                 return None
@@ -152,7 +249,12 @@ class StreamingChat(CohereObject):
             self.chatlog = response.get("chatlog")
             self.token_count = response.get("token_count")
             self.meta = response.get("meta")
-            return None
+            self.is_search_required = response.get("is_search_required")  # optional
+            self.citations = response.get("citations")  # optional
+            self.documents = response.get("documents")  # optional
+            self.search_results = response.get("search_results")  # optional
+            self.search_queries = response.get("search_queries")  # optional
+            return StreamEnd(finish_reason=finish_reason, is_finished=True, event_type=event_type, index=index)
         return None
 
     def __iter__(self) -> Generator[StreamResponse, None, None]: