+++ tests/sync/test_chat.py
@@ -3,6 +3,14 @@ import unittest
 from utils import get_api_key
 
 import cohere
+from cohere.responses.chat import (
+    StreamCitationGeneration,
+    StreamEnd,
+    StreamQueryGeneration,
+    StreamSearchResults,
+    StreamStart,
+    StreamTextGeneration,
+)
 
 API_KEY = get_api_key()
 co = cohere.Client(API_KEY)
@@ -197,3 +205,156 @@ class TestChat(unittest.TestCase):
         for logit_bias in invalid:
             with self.assertRaises(cohere.error.CohereError):
                 _ = co.chat("Yo what up?", logit_bias=logit_bias, max_tokens=5)
+
+    def test_search_queries_only_true(self):
+        prediction = co.chat(
+            "What is the height of Mount Everest? What is the depth of the Mariana Trench? What is the climate like in Nepal?",
+            search_queries_only=True,
+        )
+        self.assertTrue(prediction.is_search_required)
+        self.assertIsInstance(prediction.search_queries, list)
+        self.assertGreater(len(prediction.search_queries), 0)
+        self.assertIsInstance(prediction.search_queries[0]["text"], str)
+        self.assertIsInstance(prediction.search_queries[0]["generation_id"], str)
+
+    def test_search_queries_only_false(self):
+        prediction = co.chat("hello", search_queries_only=True)
+        self.assertFalse(prediction.is_search_required)
+        self.assertIsInstance(prediction.search_queries, list)
+        self.assertEqual(len(prediction.search_queries), 0)
+
+    def test_with_documents(self):
+        prediction = co.chat(
+            "How deep in the Mariana Trench",
+            temperature=0,
+            documents=[
+                {
+                    "id": "national_geographic_everest",
+                    "title": "Height of Mount Everest",
+                    "snippet": "The height of Mount Everest is 29,035 feet",
+                    "url": "https://education.nationalgeographic.org/resource/mount-everest/",
+                },
+                {
+                    "id": "national_geographic_mariana",
+                    "title": "Depth of the Mariana Trench",
+                    "snippet": "The depth of the Mariana Trench is 36,070 feet",
+                    "url": "https://www.nationalgeographic.org/activity/mariana-trench-deepest-place-earth",
+                },
+            ],
+        )
+        self.assertIsInstance(prediction.text, str)
+        self.assertIsInstance(prediction.citations, list)
+        self.assertGreater(len(prediction.citations), 0)
+        self.assertIsInstance(prediction.citations[0]["start"], int)
+        self.assertIsInstance(prediction.citations[0]["end"], int)
+        self.assertIsInstance(prediction.citations[0]["text"], str)
+        self.assertIsInstance(prediction.citations[0]["document_ids"], list)
+        self.assertGreater(len(prediction.citations[0]["document_ids"]), 0)
+        self.assertIsInstance(prediction.documents, list)
+        self.assertGreater(len(prediction.documents), 0)
+
+    def test_with_connectors(self):
+        prediction = co.chat("How deep in the Mariana Trench", temperature=0, connectors=[{"id": "web-search"}])
+        self.assertIsInstance(prediction.text, str)
+        self.assertIsInstance(prediction.citations, list)
+        self.assertGreater(len(prediction.citations), 0)
+        self.assertIsInstance(prediction.citations[0]["start"], int)
+        self.assertIsInstance(prediction.citations[0]["end"], int)
+        self.assertIsInstance(prediction.citations[0]["text"], str)
+        self.assertIsInstance(prediction.citations[0]["document_ids"], list)
+        self.assertGreater(len(prediction.citations[0]["document_ids"]), 0)
+        self.assertIsInstance(prediction.documents, list)
+        self.assertGreater(len(prediction.documents), 0)
+        self.assertIsInstance(prediction.search_results, list)
+        self.assertGreater(len(prediction.search_results), 0)
+
+    def test_with_citation_quality(self):
+        prediction = co.chat(
+            "How deep in the Mariana Trench",
+            citation_quality="accurate",
+            temperature=0,
+            documents=[
+                {
+                    "id": "national_geographic_mariana",
+                    "title": "Depth of the Mariana Trench",
+                    "snippet": "The depth of the Mariana Trench is 36,070 feet",
+                    "url": "https://www.nationalgeographic.org/activity/mariana-trench-deepest-place-earth",
+                },
+            ],
+        )
+        self.assertIsInstance(prediction.text, str)
+        self.assertIsInstance(prediction.citations, list)
+        self.assertGreater(len(prediction.citations), 0)
+        self.assertIsInstance(prediction.citations[0]["start"], int)
+        self.assertIsInstance(prediction.citations[0]["end"], int)
+        self.assertIsInstance(prediction.citations[0]["text"], str)
+        self.assertIsInstance(prediction.citations[0]["document_ids"], list)
+        self.assertGreater(len(prediction.citations[0]["document_ids"]), 0)
+        self.assertIsInstance(prediction.documents, list)
+        self.assertGreater(len(prediction.documents), 0)
+
+    def test_stream_with_connectors(self):
+        prediction = co.chat(
+            "How deep in the Mariana Trench", temperature=0, stream=True, connectors=[{"id": "web-search"}]
+        )
+
+        self.assertIsInstance(prediction, cohere.responses.chat.StreamingChat)
+        self.assertIsInstance(prediction.texts, list)
+        self.assertEqual(len(prediction.texts), 0)
+        self.assertIsNone(prediction.response_id)
+        self.assertIsNone(prediction.finish_reason)
+
+        expected_index = 0
+        expected_text = ""
+
+        count_stream_start = 0
+        count_text_generation = 0
+        count_query_generation = 0
+        count_citation_generation = 0
+        count_search_results = 0
+        count_stream_end = 0
+        for token in prediction:
+            if isinstance(token, StreamStart):
+                count_stream_start += 1
+                self.assertIsNotNone(token.generation_id)
+                self.assertFalse(token.is_finished)
+                self.assertEqual(token.event_type, "stream-start")
+            elif isinstance(token, StreamQueryGeneration):
+                count_query_generation += 1
+                self.assertIsNotNone(token.search_queries)
+                self.assertEqual(token.event_type, "search-queries-generation")
+            elif isinstance(token, StreamSearchResults):
+                count_search_results += 1
+                self.assertIsNotNone(token.documents)
+                self.assertIsNotNone(token.search_results)
+                self.assertEqual(token.event_type, "search-results")
+            elif isinstance(token, StreamCitationGeneration):
+                count_citation_generation += 1
+                self.assertIsNotNone(token.citations)
+                self.assertEqual(token.event_type, "citation-generation")
+            elif isinstance(token, StreamTextGeneration):
+                count_text_generation += 1
+                self.assertIsInstance(token.text, str)
+                self.assertGreater(len(token.text), 0)
+                expected_text += token.text
+                self.assertFalse(token.is_finished)
+                self.assertEqual(token.event_type, "text-generation")
+            elif isinstance(token, StreamEnd):
+                count_stream_end += 1
+                self.assertTrue(token.is_finished)
+                self.assertEqual(token.event_type, "stream-end")
+                self.assertEqual(token.finish_reason, "COMPLETE")
+            self.assertIsInstance(token.index, int)
+            self.assertEqual(token.index, expected_index)
+            expected_index += 1
+
+        self.assertEqual(count_stream_start, 1)
+        self.assertEqual(count_search_results, 1)
+        self.assertEqual(count_stream_end, 1)
+        self.assertGreaterEqual(count_citation_generation, 1)
+        self.assertGreaterEqual(count_query_generation, 1)
+        self.assertGreaterEqual(count_text_generation, 1)
+
+        self.assertEqual(prediction.texts, [expected_text])
+        self.assertIsNotNone(prediction.response_id)
+        self.assertIsNotNone(prediction.finish_reason)